Redesign tutorial for more interesting subject
Switch tutorial from one that swaps the key and the value (boring!) to one that looks for Twitter handles in tweets.
