+++ buildSrc/src/main/kotlin/common-convention.gradle.kts
@@ -17,11 +17,10 @@
 /**
  * Standard configuration of Creek aggregates
  *
- * <p>Version: 1.4
- *
  * <p>Apply to all java modules, usually excluding the root project in multi-module sets.
  *
- * <p>Version: 1.4
+ * <p>Version: 1.5
+ *  - 1.5: Add filters to exclude generated sources
  *  - 1.4: Add findsecbugs-plugin
  *  - 1.3: Fail on warnings for test code too.
  */
@@ -80,10 +79,13 @@ spotless {
         trimTrailingWhitespace()
         endWithNewline()
         toggleOffOn("formatting:off", "formatting:on")
+        targetExclude("**/build/generated/source*/**/*.*")
     }
 }
 
 spotbugs {
+    excludeFilter.set(rootProject.file("config/spotbugs/suppressions.xml"))
+
     tasks.spotbugsMain {
         reports.create("html") {
             required.set(true)
