+++ config/checkstyle/suppressions.xml
@@ -8,4 +8,6 @@
  SUPPRESSION COMMENTS OR @SuppressWarnings(..)
 -->
 <suppressions>
+    <!-- Exclude generated sources   -->
+    <suppress files="[\\/]build[\\/]generated[\\/]source" checks=".*"/>
 </suppressions>
