+++ config/spotbugs/suppressions.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<FindBugsFilter>
+    <!--
+     THIS SHOULD ONLY BE USED FOR CHECKS THAT CAN'T BE TURNED OFF WITH
+     SUPPRESSION COMMENTS OR @SuppressFBWarnings(..)
+    -->
+    <Match>
+        <!-- Exclude generated sources   -->
+        <Source name="~.*[\\/]build[\\/]generated[\\/]source.*" />
+    </Match>
+</FindBugsFilter>
