+++ docs/_demo/11-further-reading.md_
@@ -1,8 +1,3 @@
----
-title: Further reading
-permalink: /further-reading
-layout: single
----
 
 [todo]: link to additional demos, e.g.
