+++ handle-occurrence-service/Dockerfile
