+++ handle-occurrence-service/include/bin/run.sh
