--- reverse-service/src/test/resources/kafka/streams/expected_topology.txt
@@ -1,9 +1,0 @@
-Topologies:
-   Sub-topology: 0
-    Source: ingest-input (topics: [input])
-      --> switch
-    Processor: switch (stores: [])
-      --> egress-output
-      <-- ingest-input
-    Sink: egress-output (topic: output)
-      <-- switch
