+++ services/src/test/java/io/github/creek/service/basic/kafka/streams/demo/services/HandleOccurrenceServiceDescriptorTest.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2022-2023 Creek Contributors (https://github.com/creek-service)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.github.creek.service.basic.kafka.streams.demo.services;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.hasItem;
+import static org.hamcrest.Matchers.instanceOf;
+
+import java.util.List;
+import org.creekservice.api.platform.metadata.ComponentDescriptor;
+import org.creekservice.api.platform.metadata.ComponentDescriptors;
+import org.junit.jupiter.api.Test;
+
+class HandleOccurrenceServiceDescriptorTest {
+
+    @Test
+    void shouldLoadDescriptor() {
+        final List<ComponentDescriptor> loaded = ComponentDescriptors.load();
+        assertThat(loaded, hasItem(instanceOf(HandleOccurrenceServiceDescriptor.class)));
+    }
+}
