--- system-tests/src/system-test/example-suite/inputs/input.yml
@@ -1,15 +1,0 @@
-# begin-snippet: all
----
-!creek/kafka-topic@1
-topic: input
-records:
-  - key: hello
-    value: 1
-  - key: world
-    value: 2
-  - # record with explicitly null key:
-    key: ~
-    value: 3
-  - # record with implicitly null value:
-    key: x
-# end-snippet
